Bot/requirements.txt
aiofiles==24.1.0
aioshutil==1.5
asyncpraw==7.7.1
colour==0.1.5
discord.py==2.4.0
emoji==2.10.1
emojis==0.7.0
motor==3.3.2
nltk==3.8.1
profanityfilter==2.0.6
psutil==5.9.8
py7zr==0.20.8
py_cpuinfo==9.0.0
pydantic==2.9.2
pymongo==4.6.1
